add wpf version control pad
