--- src/Gui/MCSControlLibW/WLogin.xaml
@@ -1,13 +1,0 @@
-﻿<Window x:Class="MCSControlLibW.WLogin"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="MCS Control Login" Height="200" Width="400">
-    <Grid>
-        <Image HorizontalAlignment="Left" Height="101" Margin="42,34,0,0" VerticalAlignment="Top" Width="88">
-            <Image.OpacityMask>
-                <ImageBrush ImageSource="pack://siteoforigin:,,,/Resources/head.png" Stretch="None"/>
-            </Image.OpacityMask>
-        </Image>
-
-    </Grid>
-</Window>
