--- src/Gui/MCSControlW/MainWindow.xaml
@@ -1,25 +1,0 @@
-﻿<Window x:Class="MCSControlW.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Login" Height="200" Width="400" WindowStartupLocation="CenterScreen" WindowStyle="ToolWindow">
-    <Grid Margin="0,0,-8,-4">
-        <Grid.Background>
-            <LinearGradientBrush EndPoint="0.5,1" MappingMode="RelativeToBoundingBox" StartPoint="0.5,0">
-                <GradientStop Color="#FFB9E2EC" Offset="1"/>
-                <GradientStop Color="#FF206FCD"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="107*"/>
-            <ColumnDefinition Width="293*"/>
-        </Grid.ColumnDefinitions>
-        <Button x:Name="bnLogin" Content="Login" Grid.Column="1" HorizontalAlignment="Left" Margin="197,66,0,0" VerticalAlignment="Top" Width="75" Height="22" Background="{x:Null}"/>
-        <Button x:Name="bnExit" Content="Exit" Grid.Column="1" HorizontalAlignment="Left" Margin="197,116,0,0" VerticalAlignment="Top" Width="75" Height="22" IsCancel="True" Click="bnExit_Click" Background="{x:Null}"/>
-        <TextBox Grid.Column="1" HorizontalAlignment="Left" Height="23" Margin="72,65,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Grid.Column="1" HorizontalAlignment="Left" Height="23" Margin="72,115,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <Image HorizontalAlignment="Left" Height="100" Margin="20,38,0,0" VerticalAlignment="Top" Width="87"/>
-        <TextBlock Grid.Column="1" HorizontalAlignment="Left" Margin="10,68,0,0" TextWrapping="Wrap" Text="User:" VerticalAlignment="Top"/>
-        <TextBlock Grid.Column="1" HorizontalAlignment="Left" Margin="10,119,0,0" TextWrapping="Wrap" Text="Password:" VerticalAlignment="Top"/>
-
-    </Grid>
-</Window>
