+++ src/Gui/McsRemote.Control/Login.xaml
@@ -1,0 +1,8 @@
+﻿<Window x:Class="McsRemote.Control.Login"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Login" Height="300" Width="300">
+    <Grid>
+        
+    </Grid>
+</Window>
