+++ src/Gui/McsRemote.Control/Properties/Resources.resx
