+++ src/Gui/McsRemote.Control/Properties/Settings.settings
