+++ src/Gui/McsRemotePad/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="MCSControlW.App"
+﻿<Application x:Class="McsRemotePad.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="WBackForm.xaml">
+             StartupUri="MainWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
