+++ src/Gui/McsRemotePad/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace MCSControlW
+namespace McsRemotePad
 {
     /// <summary>
     /// App.xaml 的交互逻辑
