+++ src/Gui/McsRemotePad/MainWindow.xaml
@@ -1,7 +1,7 @@
-﻿<Window x:Class="MCSControlW.WMCSControl"
+﻿<Window x:Class="McsRemotePad.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="WMCSControl" Height="340.959" Width="684.133">
+        Title="MainWindow" Height="350" Width="525" Loaded="Window_Loaded_1">
     <Grid>
         
     </Grid>
