+++ src/Gui/McsRemotePad/Properties/Settings.settings
